Adding Stellar Motion Live MATLAB Script
This script loads the neccessary 'starData.mat' file and plots the chosen star's light intensity spectrum as a function of wavelengths then utilizes the hydrogen-alpha line (minimum light intensity point due to Hydrogen Absorption) to calculate the star's speed in km/s. Abstraction is added by allowing the user to modify the column number assigned to the variable 's' and make the same plots and calculations for any star found in the 'spectra' table.